requirements.txt
numpy==1.23.1
gspread==5.0.1
oauth2client==4.1.3
google-auth==2.0.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.42.0
pandas==1.4.1
scanpy==1.9.1
